# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bot Factory API — SaaS платформа для автогенерации Telegram-ботов
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /credits/purchase:
    post:
      operationId: purchaseCredits
      tags: [transactions]
      summary: Purchase a credit package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditPurchaseInput"
      responses:
        "200":
          description: Purchase initiated, credits added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditPurchaseResult"

  # ── MARKETPLACE ────────────────────────────────────────────────────────────
...
  /marketplace/{id}/deploy:
    post:
      operationId: deployTemplate
      tags: [marketplace]
      summary: Deploy a marketplace template as a new bot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateDeployInput"
      responses:
        "201":
          description: Bot created from template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"

  # ── REFERRAL ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Bot:
      type: object
      required: [id, userId, name, status, botType, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        botToken:
          type: ["string", "null"]
          description: Masked token (last 4 chars)
        botUsername:
          type: ["string", "null"]
        status:
          type: string
          enum: [generating, running, stopped, error]
        botType:
          type: string
          enum: [simple, complex, miniapp]
        generationId:
          type: ["string", "null"]
        lastActiveAt:
          type: ["string", "null"]
        createdAt:
          type: string
...
    CreditPurchaseInput:
      type: object
      required: [packageId]
      properties:
        packageId:
          type: string
          enum: [starter, basic, pro, max]
    CreditPurchaseResult:
      type: object
      required: [success, creditsAdded, newBalance]
      properties:
        success:
          type: boolean
        creditsAdded:
          type: integer
        newBalance:
          type: integer
        transactionId:
          type: string

    # ── MARKETPLACE ───────────────────────────────────────────────────────────
...
    TemplateDeployInput:
      type: object
      required: [botToken]
      properties:
        botToken:
          type: string

    # ── REFERRAL ─────────────────────────────────────────────────────────────
...