# lib/api-zod/src/generated/types/creditPurchaseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bot Factory API — SaaS платформа для автогенерации Telegram-ботов
 * OpenAPI spec version: 0.1.0
 */
import type { CreditPurchaseInputPackageId } from './creditPurchaseInputPackageId';

export interface CreditPurchaseInput {
  packageId: CreditPurchaseInputPackageId;
}