# lib/api-zod/src/generated/types/templateDeployInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bot Factory API — SaaS платформа для автогенерации Telegram-ботов
 * OpenAPI spec version: 0.1.0
 */

export interface TemplateDeployInput {
  botToken: string;
}